requirements.txt
asgiref==3.5.2
bandit==1.7.4
cffi==1.15.0
colorama==0.4.5
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
pbr==5.9.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
PyYAML==6.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
tzdata==2022.1
whitenoise==6.2.0
